+++ ui/components/combobox/base/example.jsx
@@ -57,44 +57,51 @@ export const ListboxDropdown = props => (
 ----------------------------------------------------------------------------- */
 
 // Default
-export default (
-  <Combobox
-    id={_.uniqueId('combobox-id-')}
-    aria-controls="listbox-id-1"
-    inputIconPosition="right"
-    rightInputIcon={
-      <UtilityIcon
-        symbol="down"
-        className="slds-icon slds-icon_x-small slds-icon-text-default"
-        containerClassName="slds-input__icon slds-input__icon_right"
-        assistiveText={false}
-        title={false}
+export default [
+  {
+    id: 'default',
+    label: 'Base – default (select-only)',
+    element: (
+      <Combobox
+        id={_.uniqueId('combobox-id-')}
+        aria-controls="listbox-id-1"
+        inputIconPosition="right"
+        rightInputIcon={
+          <UtilityIcon
+            symbol="down"
+            className="slds-icon slds-icon_x-small slds-icon-text-default"
+            containerClassName="slds-input__icon slds-input__icon_right"
+            assistiveText={false}
+            title={false}
+          />
+        }
+        results={
+          <Listbox
+            id="listbox-id-1"
+            snapshot={Snapshot.ObjectOptions}
+            type="plain"
+            count={8}
+            visualSelection
+          />
+        }
+        resultsType="listbox"
+        hasInteractions
+        selectOnly
       />
-    }
-    results={
-      <Listbox
-        id="listbox-id-1"
-        snapshot={Snapshot.ObjectOptions}
-        type="plain"
-        count={8}
-        visualSelection
-      />
-    }
-    resultsType="listbox"
-    hasInteractions
-  />
-);
+    )
+  }
+];
 
 // States
 export let states = [
   {
     id: 'focused',
-    label: 'Focused',
+    label: 'Focused (select-only)',
     element: (
       <Combobox
         id={_.uniqueId('combobox-id-')}
         aria-controls="listbox-id-4"
-        readonly
+        selectOnly
         inputIconPosition="right"
         rightInputIcon={
           <UtilityIcon
@@ -121,12 +128,12 @@ export let states = [
   },
   {
     id: 'focused-open',
-    label: 'Focused - Opened',
+    label: 'Focused - Opened (select-only)',
     element: (
       <Combobox
         id={_.uniqueId('combobox-id-')}
         aria-controls="listbox-id-4"
-        readonly
+        selectOnly
         inputIconPosition="right"
         rightInputIcon={
           <UtilityIcon
@@ -154,12 +161,12 @@ export let states = [
   },
   {
     id: 'highlighting-an-option',
-    label: 'Highlighting an option',
+    label: 'Highlighting an option (select-only)',
     element: (
       <Combobox
         id={_.uniqueId('combobox-id-')}
         aria-controls="listbox-id-4"
-        readonly
+        selectOnly
         inputIconPosition="right"
         rightInputIcon={
           <UtilityIcon
@@ -188,12 +195,12 @@ export let states = [
   },
   {
     id: 'selecting-a-single-option',
-    label: 'Selecting a single option',
+    label: 'Selecting a single option (select-only)',
     element: (
       <Combobox
         id={_.uniqueId('combobox-id-')}
         aria-controls="listbox-id-6"
-        readonly
+        selectOnly
         value="Accounts"
         inputIconPosition="right"
         rightInputIcon={
@@ -223,12 +230,12 @@ export let states = [
   },
   {
     id: 'selected-an-option-closed',
-    label: 'Selected an option - Closed',
+    label: 'Selected an option - Closed (select-only)',
     element: (
       <Combobox
         id={_.uniqueId('combobox-id-')}
         aria-controls="listbox-id-5"
-        readonly
+        selectOnly
         value="Accounts"
         inputIconPosition="right"
         rightInputIcon={
@@ -255,12 +262,12 @@ export let states = [
   },
   {
     id: 'selecting-multiple-options',
-    label: 'Selecting multiple options',
+    label: 'Selecting multiple options (select-only)',
     element: (
       <Combobox
         id={_.uniqueId('combobox-id-')}
         aria-controls="listbox-id-7"
-        readonly
+        selectOnly
         value="2 Options Selected"
         inputIconPosition="right"
         rightInputIcon={
@@ -289,12 +296,12 @@ export let states = [
   },
   {
     id: 'selected-multiple-options-closed',
-    label: 'Selected multiple options-closed',
+    label: 'Selected multiple options-closed (select-only)',
     element: (
       <Combobox
         id={_.uniqueId('combobox-id-')}
         aria-controls="listbox-id-8"
-        readonly
+        selectOnly
         value="2 Options Selected"
         inputIconPosition="right"
         rightInputIcon={
@@ -326,13 +333,50 @@ export let states = [
     )
   },
   {
+    id: 'selecting-a-single-option-truncated',
+    label: 'Selecting a single option (truncated)',
+    demoStyles: 'max-width: 225px;',
+    storybookStyles: true,
+    element: (
+      <Combobox
+        id={_.uniqueId('combobox-id-')}
+        aria-controls="listbox-id-6"
+        selectOnly
+        value="Accounts (should truncate)"
+        inputIconPosition="right"
+        rightInputIcon={
+          <UtilityIcon
+            symbol="down"
+            className="slds-icon slds-icon_x-small slds-icon-text-default"
+            containerClassName="slds-input__icon slds-input__icon_right"
+            assistiveText={false}
+            title={false}
+          />
+        }
+        results={
+          <Listbox
+            id="listbox-id-6"
+            snapshot={Snapshot.ObjectOptionsSingleSelection}
+            type="plain"
+            count={8}
+            visualSelection
+          />
+        }
+        resultsType="listbox"
+        aria-activedescendant="option1"
+        isOpen
+        hasFocus
+      />
+    )
+  },
+  {
     id: 'loading-options',
-    label: 'Loading more options',
+    label: 'Loading more options (select-only)',
     element: (
       <Combobox
         id={_.uniqueId('combobox-id-')}
         aria-controls="listbox-id-222"
-        readonly
+        selectOnly
         value="Accounts"
         inputIconPosition="right"
         rightInputIcon={
@@ -362,12 +406,12 @@ export let states = [
   },
   {
     id: 'grouped-options',
-    label: 'Grouped options',
+    label: 'Grouped options (select-only)',
     element: (
       <Combobox
         id={_.uniqueId('combobox-id-')}
         aria-controls="listbox-id-3"
-        readonly
+        selectOnly
         inputIconPosition="right"
         rightInputIcon={
           <UtilityIcon
@@ -403,7 +447,7 @@ export let states = [
   },
   {
     id: 'deprecated-closed',
-    label: 'Deprecated - Closed',
+    label: 'Deprecated - Closed (select-only)',
     element: (
       <div className="demo-only" style={{ height: '10rem' }}>
         <DeprecatedCombobox
@@ -417,7 +461,7 @@ export let states = [
   },
   {
     id: 'deprecated-focused',
-    label: 'Deprecated - Focused',
+    label: 'Deprecated - Focused (select-only)',
     element: (
       <div className="demo-only" style={{ height: '10rem' }}>
         <DeprecatedCombobox
@@ -432,7 +476,7 @@ export let states = [
   },
   {
     id: 'deprecated-open-item-focused',
-    label: 'Deprecated - Open - Item Focused',
+    label: 'Deprecated - Open - Item Focused (select-only)',
     element: (
       <div className="demo-only" style={{ height: '10rem' }}>
         <DeprecatedCombobox
@@ -448,7 +492,7 @@ export let states = [
   },
   {
     id: 'deprecated-closed-options-selected',
-    label: 'Deprecated - Option(s) Selected',
+    label: 'Deprecated - Option(s) Selected (select-only)',
     element: (
       <div className="demo-only" style={{ height: '10rem' }}>
         <DeprecatedCombobox