+++ ui/components/combobox/base/example.jsx
@@ -443,7 +443,6 @@ export let states = [
           <ListboxWrapper
             id="listbox-id-3"
             className="slds-dropdown slds-dropdown_fluid"
-            ariaLabel="List Items"
           >
             <ListboxGroup
               aria-label="Group One"