+++ ui/components/combobox/listbox/index.jsx
@@ -18,7 +18,6 @@ export const ListboxWrapper = props => (
     className={props.className}
     role="listbox"
     aria-orientation={props.horizontal || props.inline ? 'horizontal' : null}
-    aria-label={props.ariaLabel || 'please provide ARIA-label'}
   >
     {props.children}
   </div>
@@ -373,7 +372,6 @@ export class Listbox extends Component {
     return (
       <ListboxWrapper
         id={id}
-        ariaLabel='Menu Items'
         className={classNames(
           { 'slds-dropdown': !staticListbox },
           type === 'entity'