+++ ui/components/path/index.js
@@ -82,7 +82,6 @@ export const PathTrack = props => {
               className="slds-path__nav"
               role="listbox"
               aria-orientation="horizontal"
-              aria-label="Path Navigation"
             >
               {props.children}
             </ul>