+++ ui/components/picklist/RELEASENOTES.md
@@ -8,7 +8,7 @@
 
 ### Changed
 
-Removed the usage of the deprecated combobox. The Picklist now uses the current combobox.
+- Removed the usage of the deprecated combobox. The Picklist now uses the current combobox.
 
 ## 2.13.1